requirements.txt
# VEGA-Verified Requirements

# Core dependencies
pyyaml>=6.0
pydantic>=2.0

# SMT Solver
z3-solver>=4.12.0

# Machine Learning (optional, for neural repair model)
torch>=2.0.0
transformers>=4.30.0

# Code Parsing
tree-sitter>=0.20.0
# Note: libclang may need separate installation

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0

# Utilities
tqdm>=4.65.0
rich>=13.0.0

# Development
black>=23.0.0
isort>=5.12.0
mypy>=1.0.0
